backend/lbvitality/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
idna==2.8
numpy==1.16.2
pandas==0.24.2
psycopg2-binary==2.8.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
urllib3==1.24.2
